Add files via upload
This is the dataset that I've used in order to build the dashboard. This dataset consists of columns like name, brand, year, selling_price, km_driven, Number of Owners, seats, fuel, seller_type, transmission	engine, max_power, torque. In addition to that the dataset contains 7908 rows which makes it a vast collection of data.